pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-car-sim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "neat-python>=0.92",
    "pygame>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "hypothesis",
    "ruff",
    "mypy",
]

[project.scripts]
ai-car-sim = "ai_car_sim.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--tb=short"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
ignore_missing_imports = true
